+++ CHANGES.rst
@@ -11,6 +11,19 @@
   3.5.9, 3.6.9, 3.7.5 and 3.8.0 (final). It is also tested with PyPy2
   7.2 and PyPy 3.6 7.2
 
+- The file objects (FileObjectPosix, FileObjectThread) now
+  consistently text and binary modes. If neither 'b' nor 't' is given
+  in the mode, they will read and write native strings. If 't' is
+  given, they will always work with unicode strings, and 'b' will
+  always work with byte strings. See :issue:`1441`.
+
+- The file objects accept *encoding*, *errors* and *newline*
+  arguments. On Python 2, these are only used if 't' is in the mode.
+
+- The default mode for FileObjectPosix changed from ``rb`` to simply
+  ``r``, for consistency with the other file objects and the standard
+  ``open`` and ``io.open`` functions.
+
 
 1.5a2 (2019-10-21)
 ==================